move common code from Entity into Model
