+++ Entity/BaseTag.php
@@ -11,193 +11,14 @@
 
 namespace Sonata\NewsBundle\Entity;
 
-use Sonata\NewsBundle\Model\TagInterface;
 use Sonata\NewsBundle\Model\PostInterface;
+use Sonata\NewsBundle\Model\Tag as ModelTag;
 
-abstract class BaseTag implements TagInterface
+abstract class BaseTag extends ModelTag
 {
-    protected $name;
-
-    protected $slug;
-
-    protected $createdAt;
-
-    protected $updatedAt;
-
-    protected $enabled;
-
-    protected $posts;
-
-    /**
-     * Set name
-     *
-     * @param string $name
-     */
-    public function setName($name)
-    {
-        $this->name = $name;
-
-        $this->setSlug(self::slugify($name));
-    }
-
-    /**
-     * Get name
-     *
-     * @return string $name
-     */
-    public function getName()
-    {
-        return $this->name;
-    }
-
-    /**
-     * Set enabled
-     *
-     * @param boolean $enabled
-     */
-    public function setEnabled($enabled)
-    {
-        $this->enabled = $enabled;
-    }
-
-    /**
-     * Get enabled
-     *
-     * @return boolean $enabled
-     */
-    public function getEnabled()
-    {
-        return $this->enabled;
-    }
-
-    /**
-     * Set slug
-     *
-     * @param integer $slug
-     */
-    public function setSlug($slug)
-    {
-        $this->slug = $slug;
-    }
-
-    /**
-     * Get slug
-     *
-     * @return integer $slug
-     */
-    public function getSlug()
-    {
-        return $this->slug;
-    }
-
-    /**
-     * Set created_at
-     *
-     * @param \DateTime $createdAt
-     */
-    public function setCreatedAt(\DateTime $createdAt = null)
-    {
-        $this->createdAt = $createdAt;
-    }
-
-    /**
-     * Get created_at
-     *
-     * @return \DateTime $createdAt
-     */
-    public function getCreatedAt()
-    {
-        return $this->createdAt;
-    }
-
-    /**
-     * Set updated_at
-     *
-     * @param \DateTime $updatedAt
-     */
-    public function setUpdatedAt(\DateTime $updatedAt = null)
-    {
-        $this->updatedAt = $updatedAt;
-    }
-
-    /**
-     * Get updated_at
-     *
-     * @return datetime $updatedAt
-     */
-    public function getUpdatedAt()
-    {
-        return $this->updatedAt;
-    }
-
     public function prePersist()
     {
         $this->setCreatedAt(new \DateTime);
         $this->setUpdatedAt(new \DateTime);
-    }
-
-    public function preUpdate()
-    {
-        $this->setUpdatedAt(new \DateTime);
-    }
-
-    /**
-     * source : http://snipplr.com/view/22741/slugify-a-string-in-php/
-     *
-     * @static
-     * @param  $text
-     * @return mixed|string
-     */
-    static public function slugify($text)
-    {
-        // replace non letter or digits by -
-        $text = preg_replace('~[^\\pL\d]+~u', '-', $text);
-
-        // trim
-        $text = trim($text, '-');
-
-        // transliterate
-        if (function_exists('iconv'))
-        {
-            $text = iconv('utf-8', 'us-ascii//TRANSLIT', $text);
-        }
-
-        // lowercase
-        $text = strtolower($text);
-
-        // remove unwanted characters
-        $text = preg_replace('~[^-\w]+~', '', $text);
-
-        if (empty($text))
-        {
-            return 'n-a';
-        }
-
-        return $text;
-    }
-
-    /**
-     * Add posts
-     *
-     * @param \Sonata\NewsBundle\Model\PostInterface $posts
-     */
-    public function addPosts(PostInterface $posts)
-    {
-        $this->posts[] = $posts;
-    }
-
-    /**
-     * Get posts
-     *
-     * @return Doctrine\Common\Collections\Collection $posts
-     */
-    public function getPosts()
-    {
-        return $this->posts;
-    }
-
-    public function __toString()
-    {
-        return $this->getName();
     }
 }
