+++ Model/PostManager.php
@@ -1,0 +1,34 @@
+<?php
+
+/*
+ * This file is part of the Sonata project.
+ *
+ * (c) Thomas Rabaix <thomas.rabaix@sonata-project.org>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+namespace Sonata\NewsBundle\Model;
+
+use Sonata\NewsBundle\Model\PostManagerInterface;
+
+abstract class PostManager implements PostManagerInterface
+{
+    protected $class;
+
+     /**
+     * @return string
+     */
+    public function getClass()
+    {
+        return $this->class;
+    }
+
+    /**
+     * @return \Sonata\NewsBundle\Model\PostInterface
+     */
+    public function create()
+    {
+        return new $this->class;
+    }
+}
